Incorporated Nethtml to avoid Unix dependency.
Extracted the Nethtml parser from Ocamlnet. Removed some code that
depends on other parts of Ocamlnet that is not needed for Lambda Soup.
Also modified Nethtml to compile with stricter warnings. The LICENSE
file has been updated.

Compilation of dependencies should now be much faster due to not having
to process all of Ocamlnet. Lambda Soup is also now much more portable.

This is likely a temporary measure.

Fixes #1.