Adapt test suite to Markup.ml 0.7.6
aantron/markup.ml#30 changed the handling of </body> and </html> to be
closer to the HTML5 specification. As a result, Markup.ml no longer
generates text child nodes for the whitespace often found after </body>
and </html>.

Some of the Lambda Soup tests were expecting those text nodes to be
there, so the tests have been adjusted.