Fix the usage of type selector in :has() (#57)
The descendants function also includes text nodes,
which causes name node calls to fail down the line,
because it only expects an `Element of n, not
`Text of string, which a text node is.

In any case, :has() doesn't select text nodes. So,
excluding them beforehand is the way to go.